schema: better errors for AnyOf/AllOf + cleanups
AnyOf and AllOf have received better error messages and doc-strings.
They have also been changed so that they will accept any value if there
are no members.

Clean up mergeFieldErrors to be easer to read; Go allows appending to
nil slice and expansion of slices in append, so there is no need for
the if branch and nested for-loop that was there before.

Simplify error handling in Object.Validate slightly. Moves the ErrorMap
type from object.go to errors.go and add some helper methods.