schema: Allow AnyOf to implement FieldComparator
This changes the interface of FieldComparator so that it's possible for
FieldValidator types to determine, either run-time or compile-time, if
the validator is comparable.

AnyOf has been modified to Allow comparisons if one of it's
sub-validators allows it.

Changes the comparable types Float, Integer, AnyOf and Time to test only
the public interface; testing the private interface gives little value,
as it should be allowed to change without test updates.